Расчёт оценочной функции после каждого шага

+++ GameLogic/GameBoard.cs
@@ -215,12 +215,12 @@ namespace GameLogic
             currentShapeLocation = new ShapeLocation(binBoard[0], binBoard[1]);
         }
 
-        public int Score(ShapeLocation currentLocation)
+        public double Score(ShapeLocation currentLocation)
         {
             double dx = countCellWidth - 1 - currentShapeLocation.X;
             double dy = countCellHeight - 1 - currentShapeLocation.Y;
             double maxd = Math.Sqrt(countCellWidth * countCellWidth + countCellHeight + countCellHeight);
-            return (int)(maxd - Math.Sqrt(dx * dx + dy * dy));
+            return (maxd - Math.Sqrt(dx * dx + dy * dy));
         }
     }
 }
